+++ app/scripts/modules/cloudfoundry/src/serverGroup/configure/wizard/sections/basicSettings/BasicSettings.cf.tsx
@@ -15,7 +15,7 @@ import {
   TextInput,
 } from '@spinnaker/core';
 
-import { ICloudFoundryCreateServerGroupCommand } from '../../serverGroupConfigurationModel.cf';
+import { ICloudFoundryCreateServerGroupCommand } from 'cloudfoundry/serverGroup/configure/serverGroupConfigurationModel.cf';
 import { CloudFoundryDeploymentStrategySelector } from 'cloudfoundry/deploymentStrategy/CloudFoundryDeploymentStrategySelector';
 
 import 'cloudfoundry/common/cloudFoundry.less';
